@@ -1,7 +1,6 @@
 package io.github.trytonvanmeer.libretrivia.activities;
 
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
@@ -13,10 +12,14 @@ import io.github.trytonvanmeer.libretrivia.trivia.TriviaGame;
 public class TriviaGameResultsActivity extends BaseActivity {
     static final String EXTRA_TRIVIA_GAME = "extra_trivia_game";
 
-    @BindView(R.id.text_results_correct) TextView textResultsCorrect;
-    @BindView(R.id.text_results_wrong) TextView textResultsWrong;
-    @BindView(R.id.text_results_total) TextView textResultsTotal;
-    @BindView(R.id.button_return_to_menu) Button buttonReturnToMenu;
+    @BindView(R.id.text_results_correct)
+    TextView textResultsCorrect;
+    @BindView(R.id.text_results_wrong)
+    TextView textResultsWrong;
+    @BindView(R.id.text_results_total)
+    TextView textResultsTotal;
+    @BindView(R.id.button_return_to_menu)
+    Button buttonReturnToMenu;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -39,11 +42,6 @@ public class TriviaGameResultsActivity extends BaseActivity {
         textResultsWrong.setText(String.valueOf(game.getQuestionsCount() - correctTotal));
         textResultsTotal.setText(String.valueOf(game.getQuestionsCount()));
 
-        buttonReturnToMenu.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
+        buttonReturnToMenu.setOnClickListener(v -> finish());
     }
 }
